added plugin UI, fixed bugs with plugin installing, improved verification code, added plugin installation UI when intent comes from external app(e.g. file explorer)

+++ app/src/main/java/com/friday/ar/util/FileUtil.java
@@ -4,6 +4,7 @@ import android.util.Log;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
@@ -13,19 +14,14 @@ public class FileUtil {
 
     public static void moveFile(File f, File dest) throws IOException {
         dest.getParentFile().mkdirs();
-        FileChannel inChannel = null;
-        FileChannel outChannel = null;
-        try {
-            inChannel = new FileInputStream(f).getChannel();
-            outChannel = new FileOutputStream(dest).getChannel();
+        if (!f.exists()) {
+            throw new FileNotFoundException("The source file does not exist!\n\nsource file path:" + f.getPath());
+        }
+        try (FileChannel inChannel = new FileInputStream(f).getChannel(); FileChannel outChannel = new FileOutputStream(dest).getChannel()) {
             inChannel.transferTo(0, inChannel.size(), outChannel);
         } catch (IOException e) {
             Log.e(LOGTAG, e.getLocalizedMessage(), e);
         } finally {
-            if (inChannel != null)
-                inChannel.close();
-            if (outChannel != null)
-                outChannel.close();
             f.getParentFile().delete();
         }
     }
